[REFACTORING] Split IMAPServerAuthenticationRequireSSLTest
