+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerTest.java
@@ -261,69 +261,6 @@ class IMAPServerTest {
 
 
 
-    @Nested
-    class AuthenticationRequireSSL {
-        IMAPServer imapServer;
-
-        @AfterEach
-        void tearDown() {
-            if (imapServer != null) {
-                imapServer.destroy();
-            }
-        }
-
-        @Test
-        void loginShouldFailWhenRequireSSLAndUnEncryptedChannel() throws Exception {
-            imapServer = createImapServer("imapServerRequireSSLIsTrueAndStartSSLIsFalse.xml");
-            int port = imapServer.getListenAddresses().get(0).getPort();
-
-            assertThatThrownBy(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS))
-                .hasMessage("Login failed");
-
-        }
-
-        @Test
-        void loginShouldSuccessWhenRequireSSLAndEncryptedChannel() throws Exception {
-            imapServer = createImapServer("imapServerRequireSSLIsTrueAndStartSSLIsTrue.xml");
-            int port = imapServer.getListenAddresses().get(0).getPort();
-
-            IMAPSClient client = new IMAPSClient(false, BogusSslContextFactory.getClientContext());
-            client.setTrustManager(BogusTrustManagerFactory.getTrustManagers()[0]);
-            client.connect("127.0.0.1", port);
-            client.execTLS();
-            client.login(USER.asString(), USER_PASS);
-
-            assertThat(client.getReplyString()).contains("OK LOGIN completed.");
-        }
-
-        @Test
-        void loginShouldSuccessWhenNOTRequireSSLAndUnEncryptedChannel() throws Exception {
-            imapServer = createImapServer("imapServerRequireSSLIsFalseAndStartSSLIsFalse.xml");
-            int port = imapServer.getListenAddresses().get(0).getPort();
-
-            assertThatCode(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", SMALL_MESSAGE))
-                .doesNotThrowAnyException();
-        }
-
-        @Test
-        void loginShouldSuccessWhenNOTRequireSSLAndEncryptedChannel() throws Exception {
-            imapServer = createImapServer("imapServerRequireSSLIsFalseAndStartSSLIsTrue.xml");
-            int port = imapServer.getListenAddresses().get(0).getPort();
-
-            IMAPSClient client = new IMAPSClient(false, BogusSslContextFactory.getClientContext());
-            client.setTrustManager(BogusTrustManagerFactory.getTrustManagers()[0]);
-            client.connect("127.0.0.1", port);
-            client.execTLS();
-            client.login(USER.asString(), USER_PASS);
-
-            assertThat(client.getReplyString()).contains("OK LOGIN completed.");
-        }
-    }
 
     @Nested
     class Oidc {
